extract auth trace helpers into utils
